+++ pysui/__init__.py
@@ -43,6 +43,9 @@ from pysui.sui.sui_clients.sync_client import SuiClient as SyncClient
 from pysui.sui.sui_clients.async_client import SuiClient as AsyncClient
 
 from pysui.sui.sui_common.config import PysuiConfiguration
+from pysui.sui.sui_common.config.confgroup import GroupProtocol
+from pysui.sui.sui_common.client import PysuiClient
+from pysui.sui.sui_common.factory import client_factory
 from pysui.sui.sui_pgql.pgql_clients import SuiGQLClient as SyncGqlClient
 from pysui.sui.sui_pgql.pgql_clients import AsyncSuiGQLClient as AsyncGqlClient
 from pysui.sui.sui_grpc.pgrpc_clients import SuiGrpcClient