+++ pysui/sui/sui_pgql/pgql_types.py
@@ -1302,23 +1302,9 @@ class SystemStateSummaryGQL(PGQL_Type):
         return SystemStateSummaryGQL.from_dict(in_data)
 
 
-@dataclasses_json.dataclass_json
-@dataclasses.dataclass
-class TransactionConstraints:
-    """Subset of Protocol Constraints."""
-
-    protocol_version: Optional[int] = 0
-    max_arguments: Optional[int] = 0
-    max_input_objects: Optional[int] = 0
-    max_num_transferred_move_object_ids: Optional[int] = 0
-    max_programmable_tx_commands: Optional[int] = 0
-    max_pure_argument_size: Optional[int] = 0
-    max_tx_size_bytes: Optional[int] = 0
-    max_type_argument_depth: Optional[int] = 0
-    max_type_arguments: Optional[int] = 0
-    max_tx_gas: Optional[int] = 0
-    receive_objects: bool = False
-    enable_address_balance_gas_payments: bool = False
+# Re-exported from sui_common for backwards compatibility.
+# New code should import from pysui.sui.sui_common.sui_txn_types directly.
+from pysui.sui.sui_common.sui_txn_types import TransactionConstraints
 
 
 @dataclasses_json.dataclass_json